+++ src/components/animals/perfil/ImagesGallery.js
@@ -2,8 +2,8 @@ import React, { Component, PropTypes } from 'react';
 import Slider from 'react-slick';
 import 'slick-carousel/slick/slick.scss';
 import 'slick-carousel/slick/slick-theme.scss';
-import '../../styles/img-galeria.scss';
-import Spinner from '../common/SpinnerComponent';
+import '../../../styles/img-galeria.scss';
+import Spinner from '../../common/SpinnerComponent';
 import Lightbox from 'react-images';
 
 class ImagesGallery extends Component {
