+++ src/containers/AnimalPerfilPage.js
@@ -4,9 +4,9 @@ import { bindActionCreators } from 'redux';
 import * as animalActions from '../actions/animalActions';
 import * as confirmActions from '../actions/confirmActions';
 import * as exportActions from '../actions/exportActions';
-import InfoPerfil from '../components/animals/InfoPerfil';
-import ImagesGallery from '../components/animals/ImagesGallery';
-import AddGalleryButton from '../components/animals/AddGalleryButton';
+import InfoPerfil from '../components/animals/perfil/InfoPerfil';
+import ImagesGallery from '../components/animals/perfil/ImagesGallery';
+import AddGalleryButton from '../components/animals/perfil/AddGalleryButton';
 import AnimalEventWrapper from '../components/animals/events/AnimalEventWrapper';
 import '../styles/animal-perfil.scss';
 import { toastr } from 'react-redux-toastr';
